fix: better error messages, fewer tracebacks (#2242)
This reduces the text output when an error is hit, removing the tracebacks for known errors. Tracebacks are for unknown/unexpected errors - if you expect an error, it's better to print out an actual error message. I've also made the patch applying a bit more verbose in case it helps, as it sometimes does.

Also improvements to `BashRunner`: it no longer hangs when an error is raised and it is now a context manager