ENH Pass pywasmcross environment more carefully (#2738)
This is on top of #2734. Other build environments cannot be trusted to pass our
environment variables down #2736. We store key variables into a json file in a 
temporary directory with the symlinks. If these variables are not in `os.environ`
we load them from disk.

In addition to the added generality, this also has the advantage that it makes it 
much easier to debug specific compiler invocations from the build script because
the environment is stored to disk (that is, if we override the deletion of the
temporary directory by manually adjusting the code).