Add a JSON output for `pyodide xbuildenv search`, better tabular output (#28)
## Description

This PR closes #26. It adds a `pyodide xbuildenv search --json` option
to print a JSON-based output, along with associated tests. The advantage
is that it can be saved to a file, piped to `jq` or shell functions (or
any equivalent tools), or simply imported into a Pythonic interface.

Additionally, I added a small context manager that does not do anything
but stop printing the following lines:
```
Starting new HTTPS connection (1): raw.githubusercontent.com:443
https://raw.githubusercontent.com:443 "GET /pyodide/pyodide/main/pyodide-cross-build-environments.json HTTP/11" 200 917
```

in the output, because it conflicts with receiving valid JSON. Please
let me know if this would be undesirable. If yes, I'll try to work
around it so that it gets printed for the non-JSON output (table).