Add option to skip emscripten version check (#53)
Adds an option to skip emscripten version check. A user can either set
`SKIP_EMSCRIPTEN_VERSION_CHECK` env variable, or
`skip_emscripten_version_check` in the toml file.

I think it is useful in following scenarios:

- When we update the Emscripten version in Pyodide but it is not
released yet, so a user wants to build a package against newer
emscripten version.
- When a user needs a newer version of Emscripten to make their package
work. But the Emscripten version in Pyodide is not updated yet.

Of course, these scenarios will not work if two versions of Emscripten
are incompatible. But I think it is useful to provide this option as not
always

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Agriya Khetarpal <74401230+agriyakhetarpal@users.noreply.github.com>