Use a less sloppy check in `remove_avoided_requirements`
Resolves https://github.com/pyodide/pyodide/issues/5187. In that issue, the
build requires a package called `cmake-build-extension` but because we have
`cmake` in the list of requirements to avoid we drop it and it breaks the build.

This logic was introduced in https://github.com/pyodide/pyodide/pull/2272 and
did not come up in the review. Surprisingly this is the first problem it has
caused. Let's see if this change breaks anything in Pyodide CI.