MAINT Subclass recipe builders (#36)
Subclasses RecipeBuilder based on the package type. This is to make it
easier to distinguish between behavior that varies by package type. No
functional changes intended.

I'll update the `buildall.py` as a follow-up.

I also removed prettier from the pre-commit, which I don't think is very
useful for pyodide-build.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>