Build with wasm exception handling (#81)
For now, we need to use a custom rust toolchain for this. Apparently all
alternatives to this don't work. -Zbuild-std doesn't work with
panic=abort (https://github.com/rust-lang/cargo/issues/7359)
and my attempts to use a custom sysroot with either
https://github.com/RalfJung/rustc-build-sysroot/ or
https://github.com/DianaNites/cargo-sysroot/ seem to hit the same
problem as with `-Zbuild-std`.

Thus, I think the only reasonable way to go is to build the sysroot from
the rust source directory. This is done by
https://github.com/pyodide/rust-emscripten-wasm-eh-sysroot/