Make `pyodide venv` more compatible with standard `virtualenv` args + add tests (#117)
## Description

This PR stems from
https://github.com/pyodide/pyodide-build/issues/115#issuecomment-2706715796
where I noticed that it is currently not possible to extend `virtualenv`
args in order to override the virtual environment creation behaviour.

I've added a subset of the arguments listed at
https://virtualenv.pypa.io/en/stable/cli_interface.html. The environment
variables are already supported as we don't have much to deal with them,
this PR enables passing those arguments. A warning is raised on
currently and to-be-always unsupported arguments. For example,
`--no-pip` does not make sense as we need `pip` to be able to install
packages, and thus we always need a seeded environment.

It _might_ be possible to add support for the `--copies` and
`--symlink-app-data` flags as well, but I haven't looked into them as
they seem to be a bit more complex than I expected. Also, I'm not sure
if anyone will need the extra complexity.

Further steps that can be taken in the medium term are to create a
[`virtualenv-pyodide`
plugin](https://virtualenv.pypa.io/en/latest/extend.html) and register
it with `virtualenv` so that we can remove `venv.py` out of
`pyodide-build` altogether and maintain it separately. I would be happy
to work on this, but will probably do it in a separate repository and
shift it here.

Possible use cases:
- allow installing a custom version of `pip` when creating the virtual
environment, before it is patched up
- disable `pip` version checks when creating an environment
- embed extra packages into the environment at creation time if placed
in a local directory
- and so on

cc: @hoodmane

TODO:

- [x] Examine which args to add
- [x] Add CLI support
- [x] Documentation (most likely self-documented, so we don't need
anything else)
- [x] Add some basic tests
- [x] Figure out how to run the slow integration tests; maybe disable
isolation to let them reuse the download xbuildenv

---------

Co-authored-by: Gyeongjae Choi <def6488@gmail.com>