Add testing for macOS (#131)
## Description

This pull request closes #127 by adding macOS as a tested platform,
which should help our clients such as `cibuildwheel` and also aid local
testing for contributors on macOS devices. We could skip the integration
tests for macOS, but I think we would be fine with keeping them as we
would still be under the limit for the number of concurrent jobs on a
free GHA plan (eight). I've kept them for now.