Make needs_rebuild aware of wheel tags (#134)
This switches to using the wheel itself instead of the ".packaged"
marker to decide if a wheel needs to be rebuilt. Static and dynamic
libraries still use the ".packaged" marker.

This solves the problem that whenever I switch branches between a Python
update branch and any other branch, absolutely everything has to be
rebuilt. Instead, this saves the wheels from each branch next to each
other.

This should also resolve the annoying `Found multiple matching wheels`
errors when switching abis or Python versions.

---------

Co-authored-by: Gyeongjae Choi <def6488@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>