Use a more careful check in `remove_avoided_requirements`, and add PyWavelets recipe for testing (#60)
This PR makes the check for dependencies in `remove_avoided_requirements`
stricter in order to match the full name of the package, rather than a subset of
the string. It also adds the oldest-supported-numpy metapackage to the list
of avoided requirements, and adds PyWavelets' recipe as a test for a package
that depends on NumPy in the build graph.

Co-authored-by: Hood Chatham <roberthoodchatham@gmail.com>
Co-authored-by: Gyeongjae Choi <def6488@gmail.com>
Co-authored-by: Agriya Khetarpal <74401230+agriyakhetarpal@users.noreply.github.com>