Make create_pip_script work with uv actually this time (#185)
Weird hack to work around:
https://github.com/astral-sh/python-build-standalone/issues/380
If we resolve the symlink all the way, the python-host interpreter works
but won't install into our pyodide venv. If we don't resolve the symlink,
sys.prefix is calculated incorrectly. To ensure that we get the right
sys.prefix, we explicitly set it with the PYTHONHOME environment variable
and then call the symlink.