Install new xbuildenv if existing is for the wrong Python version (#239)
There is a confusing behavior where depending on global state, the
same command will either succeed or fail. If we run the first time
and there is no installed xbuildenv, we will install an appropriate
one. However, if we have an existing xbuildenv for the wrong Python
version, we fail.

Resolves #238.